Adds json offset, introtemplate, and outrotemplate
Some JSON sources prefix the JSON with text the prevent Cross Site Script
Inclusion (XSSI) attacks. This prefix must be stripped before passing
the text to a JSON parser.  The new json offset parameter indicates how
long the prefix that needs to be stripped is. For example, see
https://gerrit-review.googlesource.com/Documentation/rest-api.html#output.

Also adds intro template and outro template to display_external_table to
suppress the header and footer when there is no data displayed.

Change-Id: I88d849b5c8201ac03cdb3f2e085d9460a4f93bc2