Add html mode to {{#get_web_data:}}
If used with "| use xpath", DOM nodes are addressed with XPath.

Otherwise, DOM nodes are addressed with CSS/jQuery-style selectors.

Their syntax is extended with an optional .attr() tail: e.g.,
h2 a.attr('href') or h2 a.attr(href).

In both cases, HTML is parsed with DOMDocument and DOMXPath standard classes.

A new dependency is introduced: symfony/css-selector.

If it is missing, and html format without "| use xpath" is required,
i.e., with CSS/jQuery selectors, an error is shown.

In addition, the code for {{#get_web_data:}} and {{#get_file_data:}} is refactored.

Bug: T254887
Change-Id: I2704152a7c17c5c0d7d301f339f28ba70bea678e