Improve encoding detection in text functions
This commit improves encoding detection in text processing functions
like {{#get_web_data:}} and {{#get_file_data:}}.

It tries:
1) |encoding= paramter of the parser function, if any,
2) encoding in XML declaration,
3) charset in HTML <meta> tag,
4) charset in HTTP Content-Type header,
5) finally, it tries mb_detect_encoding in strict mode in the order prescribed
   by the new $edgTryEncodings configuration variable, set by default to
   ["ASCII", "UTF-8", "Windows-1251", "Windows-1252", "Windows-1254", "KOI8-R", "ISO-8859-1"].

The encoding is dealt with by the new EDUtils::detectEncodingAndConvertToUTF8() method,
which detects the encoding, changes it to UTF-8, if set in the XML/HTML explicitly
and converts the text to UTF-8.

To find the explicitly set charset, it calls EDUtils::findEncodingInText(),
also called by EDUtils::getHTMLData() in order to prepend <?xml version="1.0" encoding="UTF-8" ?>
to the HTML without explicitly declared encoding,
because DOMDocument::loadHTML() will not load correctly otherwise.

Several methods are altered in order to pass encoding and HTTP headers between them.

In addition, EDUtils::getHTMLData() is somewhat reworked to make it more robust.

Bug: T256936
Change-Id: Iecc98615f6b53b25c2761301b661df54c087a3cf