Expose five parser functions to Scribunto/Lua
Create Lua binding for five parser functions as follows:
* {{#get_web_data:}}	-> mw.ext.externaldata.getWebData
* {{#get_file_data:}}	-> mw.ext.externaldata.getFileData
* {{#get_soap_data:}}	-> mw.ext.externaldata.getSoapData
* {{#get_ldap_data:}}	-> mw.ext.externaldata.getLdapData
* {{#get_db_data:}}	-> mw.ext.externaldata.getDbData

The bindings are done in ScribuntoExternalLibraries hook,
which is only run, if Scribunto extension is enabled.

The functions accept tables as their only arguments;
in other words their arguments are effectively named. Example:

local cities = mw.ext.externaldata.getWebData {
    url = 'https://discoursedb.org/GermanyCSV.txt'
  , data = 'bordered countries=Borders,population=Population,area=Area,capital=Capital'
  , format = 'CSV with header'
}

The arguments that are comma-separated lists of assignments, like
|data=, can be supplied as tables: the above example is equivalent to:

local cities = mw.ext.externaldata.getWebData {
    url = 'https://discoursedb.org/GermanyCSV.txt'
  , data = {
        ['bordered countries'] = 'Borders'
      , population             = 'Population'
      , area                   = 'Area'
      , capital                = 'Capital'
    }
  , format = 'CSV with header'
}

All the five parser functions return their results as tables.
They do not save their output in global variables (static class
attributes after the refactoring) and therefore it cannot
be accessed with {{#external_value:}}, etc.

The tables are "row-based" rather than "column-based" as are
results of the parser functions.

If some variable has only one value (its column contains one item),
it is added to the return table as a named item, since there is a chance
that it applies to the data source as a whole, rather than to its first
row.

The example above will return:
table#1 {
 table#2 {
   ["area"] = "357,050 km²",
   ["bordered countries"] = "Austria,Belgium,Czech Republic,Denmark,France,Luxembourg,Netherlands,Poland,Switzerland",
   ["capital"] = "Berlin",
   ["population"] = "82,411,001",
 },
 ["area"] = "357,050 km²",
 ["bordered countries"] = "Austria,Belgium,Czech Republic,Denmark,France,Luxembourg,Netherlands,Poland,Switzerland",
 ["capital"] = "Berlin",
 ["population"] = "82,411,001",
}

This change required a major refactoring of code: much of it was moved
from includes/EDParserFunctions.php to includes/EDUtils.php, making
the former a kind of frontend for parserfunctions and the latter
the backend for both parser and Lua functions.

The Lua interface is provided by two new files:
* includes/EDScribunto.php and
* mw.ext.externaldata.lua.

{{#get_soap_data:}} and mw.ext.externaldata.getSoapData now really
use the cache, as is claimed by
https://www.mediawiki.org/wiki/Extension:External_Data##get_soap_data_-_web_data_via_SOAP,
but which was not the case.

In addition:
* Some global variables used to store data between function call were
  replaced with static attributes,
* One echo() call replaced with returning an error message,
* 'all' parameter added to {{#get_ldap_data:}}
  and mw.ext.externaldata.getLdapData,
* in EDHttpWithHeaders::request(), $req->execute() is wrapped
  in a try/catch block, as it can throw exceptions.

Test wiki pages (this patch already applied):
* https://traditio.wiki/Module:Test/ED for Lua functions,
* https://traditio.wiki/Project:Test/ED for parser functions.

Bug: T257715
Change-Id: I34ad35bee1e3918e4327a4cf601de0d03a336777