Add the parameter 'default xmlns prefix' when querying XML with XPath
Use this parameter as default prefix for xmlns when no prefix
is provided in the XML.

It is wiki user's responsibility to use this namespace prefix
in his XPath queries.

A working example of 'default xmlns prefix':
{{#get_web_data:
   url=https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/3028194/XML?heading=GHS+Classification
 | format=xml
 | use xpath
 | data=res=/ns:Record/ns:RecordType
 | default xmlns prefix=ns
 }}res: {{#for_external_table:{{{res}}} }}

In addition, phpcs rule MediaWiki.Commenting.RedundantVarName.Found
is disabled, so that tests may pass without disabling Doxygen.

Bug: T278661
Change-Id: If636e0fa9fa9dd1847b3d8451fff3da8aba8066b