Allow line comments in 'INI' format
Allow line comments in 'INI' format. The default comment delimiter is '#'.
This can be changed by the new setting 'comment delimiter'.
If it is set to empty string, this effectively disallows line comments.

Both kinds of comments: non key=value lines and parts of lines
starting with # and continuing to the end of the line are recorded
in the external vatiable '__comments'.

Bug: T291774
Change-Id: I82e9e78710980d75a9374a927ac19cb8bca1935e