Add throttle feature to fetchers that may require it
Add throttle feature to {{#get_web_data:}}, {{#get_soap_data:}}
and {{#get_program_data:}} and the corresponding Lua functions.

If a service has been called too recently, the call fails, producing
an error message (unless stale cache can be served, in which case
there will be a timestamp in the external variable __throttled_till),
and a job is created that purges the page, when the service frees
(with some limitations).

For {{#get_web_data:}} and {{#get_soap_data:}} throttling
is configured by two new variables:
* $edgThrottleKey (default is [ "*" => "$2nd_lvl_domain$" ])
  is an array of throttle key templates. Its keys can be:
  * hosts (i.e., domains),
  * second-level domains (e.g., livejournal.com for
    alex-mashin.livejournal.com),
  * '*' or 0 for universal fallback key.
* $edgThrottleInterval (default is [ '*' => 0 ] meaning
  no throttling) is an array of intervals between to calls
  to the same service in seconds, indexed by resolved throttle
  keys.

For any URL, $edgThrottleKey will be looked up for the complete
host, then for its second-level domain, then for the universal
fallback value. In the found template, wildcards like $host$
or $2nd_lvl_domain$ will be substituted by the corresponding
URL components or HTTP options.

The resolved throttle key will be used as index to look
for throttle interval in $edgThrottleInterval.
If it is found and non-zero, throttling will be activated.

For {{#get_program_data:}}, the throttling is configured
by $edgExeThrottleKey and $edgExeThrottleInterval, which are
arrays indexed by program ids, by default, empty.
In the found template, wildcards will be substituted with program
arguments and environment variables, although it makes
the most sense simply to set the templates constant, throttling
program calls per program itself and not any finer. Throttling
will only be activated, if both the throttling key and interval
are non-empty and non-zero.

A track of the time, when a service identified by the throttle key
is again available, is kept in ObjectCache::getLocalClusterInstance().
If this time is not yet come, the call to the throttle service
fails leaving the cache the only hope. Also, a job is created
to purge page's caches, so that the page with the error message
will be re-parsed, when it is next called.

Throttling is implemented by the trait EDConnectorThrottled
used by EDConnectorHTTP and EDConnectorExe classes.
It exports two functions:
* setupThrottle() that should be called from class constructors and
* higher-order callThrottled() that should be called from run()
  methods.

Page purge job is done by the class EDReparseJob.
It is most useful if the job queue supports delayed jobs (like Redis).

The numerous small edits to other connectors are caused by propagating
a new parameter $title (Title object) to their constructors.

Bug: T289907
Change-Id: Ia10e0676ca9996a97a1cbbfa038c1b5bc13dda48