Overcome the limitations of {{#for_external_table:}}
Currently, the use of {{#for_external_table:}} is limited
by the fact that all parser function or template calls in its input
are resolved before the local variables are substituted.

As a result, {{urlencode:{{{1}}}}} does not work as expected
and has to be replaced with {{{1:urlencode}}}, neither does
{{#if:}}.

This solution is not intuitive and universal (currently, only
two parser functions can be called this way).

The right solution would be to make the argument
to {{#for_external_table:}} behave like contents of a template,
substituting the external variables before any constructions
around them.

In this commit, this is achieved by adding a second parameter
to {{#for_external_table:}}, which behaves the required way.

Example: compare the letter case in the article titles:

old way. {{lc}} is applied to variable names:
{{#get_web_data:
  url=https://expert.ru/avtory/konstantin_pakhunov/
  |format=HTML
  |data=url=article h2>a.attr(href),title=article h2>a
}}
{| class="wikitable"
! URL !! Title {{#for_external_table:<nowiki/>
{{!}}-
{{!}} {{{url}}} {{!}}{{!}} {{lc:{{{title}}}}}
}}
|}

new way. {{lc}} is applied to variable values:
{{#get_web_data:
  url=https://expert.ru/avtory/konstantin_pakhunov/
  |format=HTML
  |data=url=article h2>a.attr(href),title=article h2>a
}}
{| class="wikitable"
! URL !! Title {{#for_external_table:|<nowiki/>
{{!}}-
{{!}} {{{url}}} {{!}}{{!}} {{lc:{{{title}}}}}
}}
|}

The difference is a pipe character after '{{#for_external_table:'.

Bug: T294704
Change-Id: Icea60067ae3353360de3f1752b4bb2728cacc776