Remove <> characters from a message
This makes a message look like it has an XML tag,
but this is not a real XML tag. For security reasons,
CI tests check message translation for code that looks
like XML tags and they sometimes need to be verified manually.
So it's better to avoid code that looks like XML tags
in messages, unless it's really needed.

Also simplify the punctuation a bit.

Change-Id: I1fa13c3d582c12ac44fc4a8ca9b890271c717790