Use HookHandlers and implement Core hook interfaces
Omit to implement the Scribuntu interface
\MediaWiki\Extension\Scribunto\Hooks\ScribuntoExternalLibrariesHook
because this whould add a dependency on the Scribunto change
If525e0f950b3a098162b19373edb3b443f559bf2 included in REL1_41.

This will be added in the follow-up change
Ie58af2ac96c635736a87e2f0f442288436860357.

The Scribunto hooks are now in a separate class to avoid a hard
dependency on extension Scribunto.

Change-Id: Ie40ab28745ca4364255b837df3ed68f0a3dc2f9f