Add MW 1.44+ compat, make macros handling more robust
Remove a strange bug (could be upstream) when if an array
is walked with &, then at the next walk with the same name
of the temporary variable, even without &, the array gets
messed up (see Addendum to
https://www.mediawiki.org/wiki/Extension_talk:External_Data#Getting_faulty_data_with_#get_web_data:_/_#get_file_data:
).

Also, add some type declarations and fix warnings about
the undefined \Title and \Html classes that caused PHAN tests to fail,
making the extension forward-compatible with MW 1.44.

Increase MW version requirement to 1.39.4.

Change-Id: I37c784ca136204c82dcd765de20727ba69ee9f2a